+++ src/main/resources/prompts/reservation.prompt
@@ -1,0 +1,30 @@
+1.Analyze User Input: Determine if the user is requesting a reservation. If reservation-related information is mentioned, extract the relevant details.
+2.Extract Required Information: Identify the following from the user's conversation:
+ - Type of Care Service: Identified by one of the following numbers:
+  - 1: Home Care
+  - 2: Daily Household Assistance
+  - Reservation Date: Extracted in the format yyyy-mm-dd.
+Reservation Time Slot: Identified by one of the following codes:
+  - A: 9:00 AM to 12:00 PM
+  - B: 12:00 PM to 3:00 PM
+  - C: 3:00 PM to 6:00 PM
+3.Check for Required Information: Ensure all necessary information is present.
+4.Request Missing Information: If any required information is missing, prompt the user for it.
+5.Return Reservation Information: If all information is provided, return it in the following JSON format:
+{
+  "actionRequired": true,  // true if all information is available
+  "serviceTypeNumber": 1,  // Type of care service number
+  "reservationDate": "2024-08-27",  // Reservation date
+  "reservationTimeCode": "A"  // Time slot code
+}
+
+Example:
+User: "I want to book Home Care for 9 AM next Tuesday."
+
+ChatGPT Result:
+{
+  "actionRequired": true,
+  "serviceTypeNumber": 1,
+  "reservationDate": "2024-08-27",
+  "reservationTimeCode": "A"
+}
