+++ src/main/resources/prompts/welfare.prompt
